Added python script and article text files
I updated the python script with the .json files that hold the article text for both putin and trump as subjects. I also attempted to upload the list of article images which are pickle files, but they were too large (55 Mb and 121 Mb, and the limit for github is 25Mb). So I will need to move these some other way directly to bridges.